load game from local storage if present
