+++ embed.go
@@ -3,4 +3,4 @@ package main
 import "embed"
 
 //go:embed data
-var data embed.FS
+var GameData embed.FS
